fix(route): add the title to the first article and refactor to V2
