+++ lib/router.js
@@ -133,8 +133,8 @@ router.get('/plainlaw/archives', lazyloadRouteHandler('./routes/plainlaw/archive
 // router.get('/jandan/:sub_model', lazyloadRouteHandler('./routes/jandan/pic'));
 
 // 喷嚏
-router.get('/dapenti/tugua', lazyloadRouteHandler('./routes/dapenti/tugua'));
-router.get('/dapenti/subject/:id', lazyloadRouteHandler('./routes/dapenti/subject'));
+// router.get('/dapenti/tugua', lazyloadRouteHandler('./routes/dapenti/tugua'));
+// router.get('/dapenti/subject/:id', lazyloadRouteHandler('./routes/dapenti/subject'));
 
 // Dockone
 router.get('/dockone/weekly', lazyloadRouteHandler('./routes/dockone/weekly'));
