--- lib/routes/dapenti/subject.js
@@ -1,5 +1,0 @@
-const utils = require('./utils');
-
-module.exports = async (ctx) => {
-    ctx.state.data = await utils.parseFeed({ subjectid: ctx.params.id });
-};
