+++ lib/v2/dapenti/maintainer.js
@@ -1,0 +1,4 @@
+module.exports = {
+    '/tugua': ['tgly307'],
+    '/subject/:id': ['xyqfer'],
+};
