+++ lib/v2/dapenti/router.js
@@ -1,0 +1,4 @@
+module.exports = function (router) {
+    router.get('/tugua', require('./tugua'));
+    router.get('/subject/:id', require('./subject'));
+};
