+++ lib/v2/dapenti/subject.js
@@ -1,0 +1,5 @@
+const utils = require('./utils');
+
+module.exports = async (ctx) => {
+    ctx.state.data = await utils.parseFeed({ subjectid: ctx.params.id }, ctx);
+};
