+++ lib/v2/dapenti/tugua.js
@@ -1,5 +1,5 @@
 const utils = require('./utils');
 
 module.exports = async (ctx) => {
-    ctx.state.data = await utils.parseFeed({ subjectid: 70 });
+    ctx.state.data = await utils.parseFeed({ subjectid: 70 }, ctx);
 };
